requirements.txt
requests
boto3
dotenv-python
fastapi
uvicorn
aiohttp